Implement header "From" rewriting
This adds header From rewriting when enabled in dma.conf with REWRITEFROM.
The original From header is added as "X-Original-From".
The maximum length of the From header is limited to 10 lines as defined
by MSG_HDR_FROM_MAX due to static variable allocation.

The rewrite function fails if the address part of the From header is
not in a single line as it uses a simple strncmp. This could be improved
by refactoring parse_addrs() into a more generic address parsing
function that returns pointers to the start and end of a valid address
in the passed string, so the rewrite function could just replace the
string between those 2 pointers (inserting line continuations before and
after the address if the address happened to be split into multiple
lines). Nevertheless, the current simplified rewriting function will
fail gracefully in this case.

If the rewrite fails to find the From address in the header line,
the header is not rewritten, but X-Original-From is still added.
This can be changed via a later commit to on failure replace it with the
envelope-from address and discard the original From line. This will ensure
rewriting always takes place, but will not preserve the From address comment
sent by the client on failure to rewrite.